feat(seed): reuse objects instead of recreate them
Currently when the seed script is re-run,
objects are removed before being recreated
even if there is no change in the payload.
That can be inconvenient when the script is used again
to repopulate global resources like workflows
because asset types or categories will be replaced,
causing reference issues.

So an optimization is now applied on asset types, workflows
and categories from seed file (not CSV).
Existing objects will be matched thanks to an alias and reused.